// src/interfaces/Response.tsx
export interface ApiResponse {
  msg: string;
  id: string;
  error?: string;
}

// Generated by https://quicktype.io

export interface GetDataAppResponse {
  data: DataApp[];
}

export interface DataApp {
  _id:   string;
  title: string;
  body:  string;
  image: string;
}

export interface GetDataCardsResponse {
  data: DataCard[];
}

export interface DataCard {
  _id:          string;
  title:        string;
  image:        string;
  path:         string;
  body:         string;
  imageDetails: string;
}

export interface GetDataMaterialsResponse {
  data: DataMaterials[];
}

export interface DataMaterials {
  _id:   string;
  title: string;
  items: Item[];
}

export interface Item {
  _id:   string;
  title: string;
  image: string;
  href:  string;
}

export interface GetDataResponse {
  error?: string
  data: DataApp[] | DataCard[] | DataMaterials[]
}

export interface GetFormsResponse {
  forms: Form[];
}

export interface Form {
  extensionFile: string;
  originalname:  string;
  _id:           string;
  first_name:    string;
  last_name:     string;
  msg:           string;
  email:         string;
  file:          string;
  url:           string;
  createdAt:     string;
  updatedAt:     string;
  __v:           number;
}